Fixed the issues in Controller/DiscordEvents.js
Well, there were a couple of issues in this file and the XenProvider.

The issue here was that on line 69.
**operationData.edited_date === event.operation_edited_date**
was always equal to each other since both dates ended up being the Games' last Edited date and not the Operation. I learned about this after switching the test operation between 2 different games, which triggered the update function.

The other issue was on lines 165-166 and 175-176. Discord has switched from string to numbers when specifying the type. So that resulted in a badly formatted body upon submittal.